Refactor: Split main.go into logical modules and add proper tests
